Automata map generator update
I mark the regions and as I'm marking I also compute each of the
regions' area. The region with the largest area remains, others become
walls. At the same time, if the region area is less than half the map's
area, I invert the map, that is, walls become floors, floors become
walls and I get the larger area as the floors. Check the comments as
well.